Added Hover Style
Added Hover Style to the Search page list item labels, the hover effects makes the page feel a bit more responsive. Without it, it didnt feel as apparent what the lis were actually doing. That being said, on mobile it looks tacky, so a media query may be best here